Move the table definitions for the connection pool tests out of the code into the database-specific SQL scripts and generated models..
