+++ tests/core/module/sqlpp23.test.core.tables.cppm
@@ -1,7 +1,7 @@
 module;
 
 // clang-format off
-// generated by ./scripts/sqlpp23-ddl2cpp --path-to-ddl tests/include/sqlpp23/tests/core/tab_department.sql tests/include/sqlpp23/tests/core/tables.sql --path-to-module tests/core/module/sqlpp23.test.core.tables.cppm --module-name=sqlpp23.test.core.tables --namespace test --assume-auto-id
+// generated by ./scripts/sqlpp23-ddl2cpp --path-to-ddl tests/include/sqlpp23/tests/core/tables.sql --path-to-module tests/core/module/sqlpp23.test.core.tables.cppm --module-name=sqlpp23.test.core.tables --namespace test --assume-auto-id
 
 #include <optional>
 
@@ -12,32 +12,6 @@ import sqlpp23.core;
 export module sqlpp23.test.core.tables;
 
 namespace test {
-  export struct TabDepartment_ {
-    struct Id {
-      SQLPP_CREATE_NAME_TAG_FOR_SQL_AND_CPP(id, id);
-      using data_type = ::sqlpp::integral;
-      using has_default = std::true_type;
-    };
-    struct Name {
-      SQLPP_CREATE_NAME_TAG_FOR_SQL_AND_CPP(name, name);
-      using data_type = std::optional<::sqlpp::text>;
-      using has_default = std::true_type;
-    };
-    struct Division {
-      SQLPP_CREATE_NAME_TAG_FOR_SQL_AND_CPP(division, division);
-      using data_type = ::sqlpp::text;
-      using has_default = std::true_type;
-    };
-    SQLPP_CREATE_NAME_TAG_FOR_SQL_AND_CPP(tab_department, tabDepartment);
-    template<typename T>
-    using _table_columns = sqlpp::table_columns<T,
-               Id,
-               Name,
-               Division>;
-    using _required_insert_columns = sqlpp::detail::type_set<>;
-  };
-  export using TabDepartment = ::sqlpp::table_t<TabDepartment_>;
-
   export struct TabFoo_ {
     struct Id {
       SQLPP_CREATE_NAME_TAG_FOR_SQL_AND_CPP(id, id);
