--- tests/include/sqlpp23/tests/core/tab_department.sql
@@ -1,6 +1,0 @@
-CREATE TABLE tab_department (
-   id INTEGER PRIMARY KEY AUTO_INCREMENT,
-   name CHAR(100),
-   division VARCHAR(255) NOT NULL DEFAULT 'engineering'
-);
-
