+++ tests/mysql/usage/ConnectionPool.cpp
@@ -36,11 +36,6 @@ int ConnectionPool(int, char*[]) {
   try {
     sqlpp::test::test_connection_pool<sql::connection_pool>(
         sql::make_test_config(),
-        "CREATE TABLE tab_department ("
-        "id INTEGER PRIMARY KEY AUTO_INCREMENT, "
-        "name CHAR(100), "
-        "division VARCHAR(255) NOT NULL DEFAULT 'engineering'"
-        ")",
         mysql_thread_safe());
   } catch (const std::exception& e) {
     std::cerr << "Exception: " << e.what() << std::endl;
